requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
Pillow==10.2.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
razorpay==1.4.2
